refactor(netlist): extract the JSON serializer into netlist_json.py
The netlist_to_dict cluster (18 _*_to_dict functions) moves to a new
self-contained netlist_json.py (448 lines) -- it imports only model types plus
the graph health/properties serializers, nothing from netlist/render_lvnet.
netlist.py re-exports netlist_to_dict so all call sites (cli, mcp, 7 tests) are
unchanged. netlist.py 4,121 -> 3,717.

Byte-identical lvnet; JSON output unchanged (test_netlist_json green); pyright
src/ clean; no import cycle.
